+++ src/lib.rs
@@ -5,12 +5,12 @@ mod encoding;
 mod poll;
 mod driver;
 mod engine;
+mod adapters;
+mod remote_addr;
 
-pub mod remote_addr;
 pub mod adapter;
-pub mod adapters;
 pub mod network;
 pub mod events;
 
 pub use adapters::udp::MAX_UDP_PAYLOAD_LEN;
-//pub use adapters::web_socket::MAX_WS_PAYLOAD_LEN;
+pub use adapters::web_socket::MAX_WS_PAYLOAD_LEN;